fix: update `landing-zone-vsi` module to v3.0.0 (#638)
BREAKING CHANGE: If your SLZ solution provisions a VSI and you have enabled a floating IP for it, when upgrading to this version floating IPs are going to be deleted and recreated upon apply. This is due to a bug where the floating IPs were being created incorrectly in the Default resource group. Upon re-creation, the floating IPs will be created in the same resource group as the VSI.

Please plan accordingly before upgrading incase this change will cause disruption for whatever is using the floating IPs.

NOTE: By default, the only deployable architecture impacted here is the VSI quickstart flavor, as it provisions a floating IP for use as a jumpbox. If however you have customised any of the other deployable architectures to provision any floating IPs, they will also be impacted.