feat: updated [landing-zone-vsi](https://github.com/terraform-ibm-modules/terraform-ibm-landing-zone-vsi) module to `4.2.0`<br>- Refactored the logic used to create OCP clusters. Its now using the [base-ocp-vpc](https://github.com/terraform-ibm-modules/terraform-ibm-base-ocp-vpc) module<br>- This module has some extra functionality which requires the runtime to have access to IBM Cloud private endpoints. (#870)
BREAKING CHANGE: If you were to upgrade the 'Red Hat OpenShift Container Platform on VPC landing zone' solution, or the 'VSI on VPC landing zone' from a previous version to this version, infrastructure will be destroyed and recreated. A fully supported migration method will be available shortly. If re-creating infrastructure is going to impact day-to-day operations, we suggest to hold off upgrading to this version until there is a fully supported migration path.